Move qggmrf functions to separate file.
