+++ mbirjax/__init__.py
@@ -1,6 +1,7 @@
 from .projectors import *
 from .parameter_handler import *
 from .tomography_model import *
+from .qggmrf import *
 from .parallel_beam import *
 from .cone_beam import *
 from .vcd_utils import *
